Introduce "imageInsert" plugin as wrapper for "imageUpload".
